Consolidate validation
