+++ src/main/java/com/amannmalik/mcp/elicitation/ElicitRequest.java
@@ -37,9 +37,9 @@ public record ElicitRequest(String message, JsonObject requestedSchema, JsonObje
             throw new IllegalArgumentException("message and requestedSchema are required");
         }
         this.message = InputSanitizer.requireClean(message);
-        ElicitationSchemaValidator.requireValid(requestedSchema);
+        ValidationUtil.requireElicitSchema(requestedSchema);
         this.requestedSchema = requestedSchema;
-        MetaValidator.requireValid(_meta);
+        ValidationUtil.requireMeta(_meta);
         this._meta = _meta;
     }
 }
