+++ src/main/java/com/amannmalik/mcp/elicitation/ElicitResult.java
@@ -3,7 +3,7 @@ package com.amannmalik.mcp.elicitation;
 import com.amannmalik.mcp.core.JsonCodec;
 import com.amannmalik.mcp.core.AbstractEntityCodec;
 import com.amannmalik.mcp.validation.InputSanitizer;
-import com.amannmalik.mcp.validation.MetaValidator;
+import com.amannmalik.mcp.validation.ValidationUtil;
 import jakarta.json.*;
 
 import java.util.Set;
@@ -62,6 +62,6 @@ public record ElicitResult(ElicitationAction action, JsonObject content, JsonObj
                 }
             }
         }
-        MetaValidator.requireValid(_meta);
+        ValidationUtil.requireMeta(_meta);
     }
 }
