+++ src/main/java/com/amannmalik/mcp/ping/PingRequest.java
@@ -2,7 +2,7 @@ package com.amannmalik.mcp.ping;
 
 import com.amannmalik.mcp.core.AbstractEntityCodec;
 import com.amannmalik.mcp.core.JsonCodec;
-import com.amannmalik.mcp.validation.MetaValidator;
+import com.amannmalik.mcp.validation.ValidationUtil;
 import jakarta.json.JsonObject;
 
 public record PingRequest(JsonObject _meta) {
@@ -10,6 +10,6 @@ public record PingRequest(JsonObject _meta) {
             AbstractEntityCodec.metaOnly(PingRequest::_meta, PingRequest::new);
 
     public PingRequest {
-        MetaValidator.requireValid(_meta);
+        ValidationUtil.requireMeta(_meta);
     }
 }
