+++ src/main/java/com/amannmalik/mcp/prompts/PromptArgument.java
@@ -4,7 +4,7 @@ import com.amannmalik.mcp.core.AbstractEntityCodec;
 import com.amannmalik.mcp.core.JsonCodec;
 import com.amannmalik.mcp.util.DisplayNameProvider;
 import com.amannmalik.mcp.validation.InputSanitizer;
-import com.amannmalik.mcp.validation.MetaValidator;
+import com.amannmalik.mcp.validation.ValidationUtil;
 import jakarta.json.*;
 
 import java.util.Set;
@@ -43,7 +43,7 @@ public record PromptArgument(
         name = InputSanitizer.requireClean(name);
         title = InputSanitizer.cleanNullable(title);
         description = InputSanitizer.cleanNullable(description);
-        MetaValidator.requireValid(_meta);
+        ValidationUtil.requireMeta(_meta);
     }
 
 }
