+++ src/main/java/com/amannmalik/mcp/resources/ResourceUpdate.java
@@ -1,11 +1,11 @@
 package com.amannmalik.mcp.resources;
 
 import com.amannmalik.mcp.validation.InputSanitizer;
-import com.amannmalik.mcp.validation.UriValidator;
+import com.amannmalik.mcp.validation.ValidationUtil;
 
 public record ResourceUpdate(String uri, String title) {
     public ResourceUpdate {
-        uri = UriValidator.requireAbsolute(uri);
+        uri = ValidationUtil.requireAbsoluteUri(uri);
         title = InputSanitizer.cleanNullable(title);
     }
 }
