+++ src/main/java/com/amannmalik/mcp/roots/ListRootsRequest.java
@@ -2,7 +2,7 @@ package com.amannmalik.mcp.roots;
 
 import com.amannmalik.mcp.core.AbstractEntityCodec;
 import com.amannmalik.mcp.core.JsonCodec;
-import com.amannmalik.mcp.validation.MetaValidator;
+import com.amannmalik.mcp.validation.ValidationUtil;
 import jakarta.json.JsonObject;
 
 public record ListRootsRequest(JsonObject _meta) {
@@ -10,6 +10,6 @@ public record ListRootsRequest(JsonObject _meta) {
             AbstractEntityCodec.metaOnly(ListRootsRequest::_meta, ListRootsRequest::new);
 
     public ListRootsRequest {
-        MetaValidator.requireValid(_meta);
+        ValidationUtil.requireMeta(_meta);
     }
 }
