+++ src/main/java/com/amannmalik/mcp/tools/CallToolRequest.java
@@ -3,7 +3,7 @@ package com.amannmalik.mcp.tools;
 import com.amannmalik.mcp.core.AbstractEntityCodec;
 import com.amannmalik.mcp.core.JsonCodec;
 import com.amannmalik.mcp.validation.InputSanitizer;
-import com.amannmalik.mcp.validation.MetaValidator;
+import com.amannmalik.mcp.validation.ValidationUtil;
 import jakarta.json.*;
 
 import java.util.Set;
@@ -36,6 +36,6 @@ public record CallToolRequest(String name,
         if (name == null) throw new IllegalArgumentException("name required");
         name = InputSanitizer.requireClean(name);
         arguments = arguments == null ? JsonValue.EMPTY_JSON_OBJECT : arguments;
-        MetaValidator.requireValid(_meta);
+        ValidationUtil.requireMeta(_meta);
     }
 }
